requirements.txt
openpyxl=2.5.4
python-docx=0.8.6